requirements.txt
flask
sqlalchemy
python-dotenv
openai
